requirements.txt
PyYAML
pandas
plotly